adk-travel-assistant/requirements.txt
google-cloud-aiplatform==1.38.1
flask==3.0.0
google-cloud-bigquery==3.15.0
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.0 